TM4J-13478 Move tool to scoped directory
